+++ TestSupport/EfSchemeCompare/CompareEfSqlConfig.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Linq;
 
 namespace TestSupport.EfSchemeCompare
 {
@@ -13,13 +12,14 @@ namespace TestSupport.EfSchemeCompare
     /// </summary>
     public class CompareEfSqlConfig
     {
-        private List<CompareLog> _logsToIgnore = new List<CompareLog>();
+        private readonly List<CompareLog> _logsToIgnore = new List<CompareLog>();
 
         /// <summary>
-        /// Place a table name (with schema name if required) of the tables you want the comparer to ignore.
-        /// This allows you to ignore tables that your EF Core context doesn't use. 
-        /// This will stop "Extra In Database" errors for these tables
-        /// Typical format: "MyTable,MyOtherTable,dbo.MyTableWithSchema" (note: table match is case insensitive)
+        /// This allows you to ignore tables that your EF Core context doesn't use. There are three settings
+        /// 1. null - this will only check the tables that the DbContexts are mapped to.
+        /// 2. "" - This will check all tables in the database against the entity classes in the DbContexts.
+        /// 3. A comma delimited list of tables, with optional schema, to ignore. 
+        ///    Typical format: "MyTable,MyOtherTable,MySchema.MyTableWithSchema" (note: the schema/table match is case insensitive)
         /// </summary>
         public string TablesToIgnoreCommaDelimited { get; set; }