+++ ReleaseNotes.md
@@ -2,16 +2,15 @@
 
 ## TODO
 
-- Bug fix: Fix issus #15 - default SQL value not checked properly.
 - Improvement: Add case insensitive table, schema, columns, etc. matching feature to EfSchemaCompare. Issues #9 and #19.
 - Think about better IgnoreErrors...
 - Update EfSchemaCompare docs about TablesToIgnoreCommaDelimited
-- Double-check the default deserialise settings in Seed to Production
 
 ## 2.0.0
 
 - Improvement (BREAKING CHANGE): By default EfSchemaCompare to only scan the tables that the entity classes map to - see issue #18.
 - Bug fix: Fixed problem with serializing/deserializing DDD-styled entity classes.
+- Bug fix: Default SQL value handled improved, plus now sets the correct ValueGenerated - see issue #15
 
 ## 1.9.0